chore(dev): conditional workflow steps
