+++ .gitignore
@@ -200,3 +200,5 @@ rabbitmq_data
 
 backend/.test_token_compression.py
 backend/test_token_compression_data.py
+
+dump.rdb
