+++ frontend/src/components/agents/workflows/conditional-workflow-builder.tsx
@@ -4,31 +4,33 @@ import React, { useState, useCallback, useMemo } from 'react';
 import { 
   Plus, 
   Trash2, 
-  GitBranch,
+  AlertTriangle,
   ChevronsUpDown,
   Check,
-  Zap
+  MoreHorizontal
 } from 'lucide-react';
 import { Button } from '@/components/ui/button';
-import { Input } from '@/components/ui/input';
+
 import { Label } from '@/components/ui/label';
-import { Card } from '@/components/ui/card';
 import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover';
 import { Command, CommandEmpty, CommandGroup, CommandInput, CommandItem, CommandList } from '@/components/ui/command';
 import { cn } from '@/lib/utils';
+import { Input } from '@/components/ui/input';
 
 export interface ConditionalStep {
   id: string;
   name: string;
   description: string;
-  type: 'instruction' | 'condition' | 'parallel';
+  type: 'instruction' | 'condition' | 'parallel' | 'sequence';
   config: Record<string, any>;
   conditions?: {
-    type: 'if' | 'else';
-    expression?: string; // Natural language condition like "if user asks about pricing"
+    type: 'if' | 'else' | 'elseif';
+    expression?: string;
   };
-  children?: ConditionalStep[]; // For branches
+  children?: ConditionalStep[];
   order: number;
+  enabled?: boolean;
+  hasIssues?: boolean;
 }
 
 interface ConditionalWorkflowBuilderProps {
@@ -69,22 +71,22 @@ export function ConditionalWorkflowBuilder({
   isLoadingTools 
 }: ConditionalWorkflowBuilderProps) {
   const [toolSearchOpen, setToolSearchOpen] = useState<{[key: string]: boolean}>({});
+  const [activeConditionTab, setActiveConditionTab] = useState<{[key: string]: string}>({});
 
   const generateId = () => Math.random().toString(36).substr(2, 9);
 
   const addStep = useCallback((parentId?: string, afterStepId?: string) => {
     const newStep: ConditionalStep = {
       id: generateId(),
-      name: 'New Step',
+      name: 'Step',
       description: '',
       type: 'instruction',
       config: {},
-      order: 0
+      order: 0,
+      enabled: true,
     };
-
     const updateSteps = (items: ConditionalStep[]): ConditionalStep[] => {
       if (!parentId) {
-        // Add to root level
         if (afterStepId) {
           const index = items.findIndex(s => s.id === afterStepId);
           return [...items.slice(0, index + 1), newStep, ...items.slice(index + 1)];
@@ -92,7 +94,6 @@ export function ConditionalWorkflowBuilder({
         return [...items, newStep];
       }
 
-      // Add as child of parent
       return items.map(step => {
         if (step.id === parentId) {
           return {
@@ -116,24 +117,15 @@ export function ConditionalWorkflowBuilder({
   const addCondition = useCallback((afterStepId: string) => {
     const ifStep: ConditionalStep = {
       id: generateId(),
-      name: 'If condition',
+      name: 'If',
       description: '',
       type: 'condition',
       config: {},
       conditions: { type: 'if', expression: '' },
       children: [],
-      order: 0
-    };
-
-    const elseStep: ConditionalStep = {
-      id: generateId(),
-      name: 'Otherwise',
-      description: '',
-      type: 'condition',
-      config: {},
-      conditions: { type: 'else' },
-      children: [],
-      order: 1
+      order: 0,
+      enabled: true,
+      hasIssues: true
     };
 
     const updateSteps = (items: ConditionalStep[]): ConditionalStep[] => {
@@ -142,12 +134,86 @@ export function ConditionalWorkflowBuilder({
         return [
           ...items.slice(0, index + 1),
           ifStep,
+          ...items.slice(index + 1)
+        ];
+      }
+      
+      return items.map(step => {
+        if (step.children) {
+          return {
+            ...step,
+            children: updateSteps(step.children)
+          };
+        }
+        return step;
+      });
+    };
+
+    onStepsChange(updateSteps(steps));
+  }, [steps, onStepsChange]);
+
+  const addElseCondition = useCallback((siblingId: string) => {
+    const elseIfStep: ConditionalStep = {
+      id: generateId(),
+      name: 'Else If',
+      description: '',
+      type: 'condition',
+      config: {},
+      conditions: { type: 'elseif', expression: '' },
+      children: [],
+      order: 0,
+      enabled: true,
+      hasIssues: true
+    };
+
+    const updateSteps = (items: ConditionalStep[]): ConditionalStep[] => {
+      const index = items.findIndex(s => s.id === siblingId);
+      if (index !== -1) {
+        return [
+          ...items.slice(0, index + 1),
+          elseIfStep,
+          ...items.slice(index + 1)
+        ];
+      }
+      
+      return items.map(step => {
+        if (step.children) {
+          return {
+            ...step,
+            children: updateSteps(step.children)
+          };
+        }
+        return step;
+      });
+    };
+
+    onStepsChange(updateSteps(steps));
+  }, [steps, onStepsChange]);
+
+  const addFinalElse = useCallback((siblingId: string) => {
+    const elseStep: ConditionalStep = {
+      id: generateId(),
+      name: 'Else',
+      description: '',
+      type: 'condition',
+      config: {},
+      conditions: { type: 'else' },
+      children: [],
+      order: 0,
+      enabled: true,
+      hasIssues: false
+    };
+
+    const updateSteps = (items: ConditionalStep[]): ConditionalStep[] => {
+      const index = items.findIndex(s => s.id === siblingId);
+      if (index !== -1) {
+        return [
+          ...items.slice(0, index + 1),
           elseStep,
           ...items.slice(index + 1)
         ];
       }
       
-      // Check in children
       return items.map(step => {
         if (step.children) {
           return {
@@ -166,7 +232,17 @@ export function ConditionalWorkflowBuilder({
     const updateSteps = (items: ConditionalStep[]): ConditionalStep[] => {
       return items.map(step => {
         if (step.id === stepId) {
-          return { ...step, ...updates };
+          const updatedStep = { ...step, ...updates };
+          if (updatedStep.type === 'instruction' && updatedStep.name && updatedStep.name !== 'New Step') {
+            updatedStep.hasIssues = false;
+          } else if (updatedStep.type === 'condition' && 
+                    (updatedStep.conditions?.type === 'if' || updatedStep.conditions?.type === 'elseif') && 
+                    updatedStep.conditions?.expression) {
+            updatedStep.hasIssues = false;
+          } else if (updatedStep.type === 'condition' && updatedStep.conditions?.type === 'else') {
+            updatedStep.hasIssues = false;
+          }
+          return updatedStep;
         }
         if (step.children) {
           return {
@@ -177,7 +253,6 @@ export function ConditionalWorkflowBuilder({
         return step;
       });
     };
-
     onStepsChange(updateSteps(steps));
   }, [steps, onStepsChange]);
 
@@ -199,286 +274,414 @@ export function ConditionalWorkflowBuilder({
     onStepsChange(removeFromSteps(steps));
   }, [steps, onStepsChange]);
 
-  const renderStep = (step: ConditionalStep, depth: number = 0, isLast: boolean = false) => {
-    const isCondition = step.type === 'condition';
-    const hasChildren = step.children && step.children.length > 0;
+  const getStepNumber = useCallback((stepId: string, items: ConditionalStep[] = steps, counter = { value: 0 }): number => {
+    for (const step of items) {
+      counter.value++;
+      if (step.id === stepId) {
+        return counter.value;
+      }
+      if (step.children && step.children.length > 0) {
+        const found = getStepNumber(stepId, step.children, counter);
+        if (found > 0) return found;
+      }
+    }
+    return 0;
+  }, [steps]);
+
+  const getConditionLetter = (index: number) => {
+    return String.fromCharCode(65 + index);
+  };
+
+  const renderConditionTabs = (conditionSteps: ConditionalStep[], groupKey: string) => {
+    const activeTabId = activeConditionTab[groupKey] || conditionSteps[0]?.id;
+    const activeStep = conditionSteps.find(s => s.id === activeTabId) || conditionSteps[0];
+    const hasElse = conditionSteps.some(step => step.conditions?.type === 'else');
+
+    const handleKeyDown = (e: React.KeyboardEvent, step: ConditionalStep) => {
+      if (e.key === 'Backspace' || e.key === 'Delete') {
+        e.preventDefault();
+        if (conditionSteps.length > 1 && !(conditionSteps.length === 1 && step.conditions?.type === 'if')) {
+          removeStep(step.id);
+          const remainingConditions = conditionSteps.filter(s => s.id !== step.id);
+          if (remainingConditions.length > 0) {
+            setActiveConditionTab(prev => ({ ...prev, [groupKey]: remainingConditions[0].id }));
+          }
+        }
+      }
+    };
 
     return (
-      <div key={step.id} className="relative">
-        <div className={cn(
-          "flex items-start gap-2",
-          depth > 0 && "ml-10"
-        )}>
-          {depth > 0 && (
-            <div className="absolute -left-10 top-4 flex items-center">
-              <div className="w-8 h-px bg-border" />
-              <div className="w-2 h-2 rounded-full bg-border" />
-            </div>
-          )}
-          
-          <div className={cn(
-            "w-8 h-8 rounded-lg flex items-center justify-center shrink-0 mt-0.5",
-            isCondition 
-              ? "bg-violet-500/10 text-violet-500 dark:bg-violet-500/20" 
-              : "bg-primary/10 text-primary dark:bg-primary/20"
-          )}>
-            {isCondition ? <GitBranch className="h-4 w-4" /> : <Zap className="h-4 w-4" />}
-          </div>
-
-          <div className="flex-1 min-w-0">
-            <Card className={cn(
-              "transition-all p-0",
-              isCondition 
-                ? "border-violet-500/20 bg-violet-50/50 dark:bg-violet-950/20 dark:border-violet-500/30" 
-                : "border-border hover:border-primary/20"
-            )}>
-              <div className="p-3 space-y-2">
-                {isCondition && step.conditions?.type === 'if' ? (
-                  <div className="space-y-1.5">
-                    <Label className="text-xs font-medium text-violet-600 dark:text-violet-400">When this is true:</Label>
-                    <Input
-                      value={step.conditions.expression || ''}
-                      onChange={(e) => updateStep(step.id, { 
-                        conditions: { ...step.conditions, expression: e.target.value } 
-                      })}
-                      placeholder="e.g., user asks about pricing"
-                      className="h-8 text-sm bg-background/50 border-violet-200 dark:border-violet-800 placeholder:text-muted-foreground/60"
-                    />
-                  </div>
-                ) : isCondition && step.conditions?.type === 'else' ? (
-                  <div className="text-sm font-semibold text-violet-600 dark:text-violet-400">Otherwise</div>
-                ) : (
-                  <>
-                    <Input
-                      value={step.name}
-                      onChange={(e) => updateStep(step.id, { name: e.target.value })}
-                      placeholder="Step name"
-                      className="border-0 bg-transparent shadow-none text-sm font-semibold p-0 h-auto focus-visible:ring-0 placeholder:text-muted-foreground"
-                    />
-                    <Input
-                      value={step.description}
-                      onChange={(e) => updateStep(step.id, { description: e.target.value })}
-                      placeholder="What does this step do?"
-                      className="border-0 bg-transparent shadow-none text-xs text-muted-foreground p-0 h-auto focus-visible:ring-0 placeholder:text-muted-foreground/60"
-                    />
-                    
-                    {!isCondition && (
-                      <Popover 
-                        open={toolSearchOpen[step.id] || false} 
-                        onOpenChange={(open) => setToolSearchOpen(prev => ({ ...prev, [step.id]: open }))}
-                      >
-                        <PopoverTrigger asChild>
-                          <Button
-                            variant="outline"
-                            role="combobox"
-                            aria-expanded={toolSearchOpen[step.id] || false}
-                            className="h-7 w-full justify-between text-xs bg-background/50 border-muted-foreground/20 hover:border-muted-foreground/40"
-                          >
-                            {step.config.tool_name ? (
-                              <span className="flex items-center gap-1.5 text-xs">
-                                {(() => {
-                                  const agentpressTool = agentTools?.agentpress_tools.find(t => t.name === step.config.tool_name);
-                                  if (agentpressTool) {
-                                    return (
-                                      <>
-                                        <span>{agentpressTool.icon || '🔧'}</span>
-                                        <span>{normalizeToolName(agentpressTool.name, 'agentpress')}</span>
-                                      </>
-                                    );
-                                  }
-                                  const mcpTool = agentTools?.mcp_tools.find(t => `${t.server}:${t.name}` === step.config.tool_name);
-                                  if (mcpTool) {
-                                    return (
-                                      <>
-                                        <span>{mcpTool.icon || '🔧'}</span>
-                                        <span>{normalizeToolName(mcpTool.name, 'mcp')}</span>
-                                      </>
-                                    );
-                                  }
-                                  return step.config.tool_name;
-                                })()}
-                              </span>
-                            ) : (
-                              <span className="text-muted-foreground">Select tool (optional)</span>
-                            )}
-                            <ChevronsUpDown className="ml-2 h-3 w-3 shrink-0 opacity-50" />
-                          </Button>
-                        </PopoverTrigger>
-                        <PopoverContent className="w-[280px] p-0" align="start">
-                          <Command>
-                            <CommandInput placeholder="Search tools..." className="h-8 text-xs" />
-                            <CommandEmpty className="py-2 text-xs text-center">No tools found.</CommandEmpty>
-                            <CommandList>
-                              {isLoadingTools ? (
-                                <CommandItem disabled className="text-xs">Loading tools...</CommandItem>
-                              ) : agentTools ? (
-                                <>
-                                  {agentTools.agentpress_tools.filter(tool => tool.enabled).length > 0 && (
-                                    <CommandGroup heading="AgentPress Tools">
-                                      {agentTools.agentpress_tools.filter(tool => tool.enabled).map((tool) => (
-                                        <CommandItem
-                                          key={tool.name}
-                                          value={`${normalizeToolName(tool.name, 'agentpress')} ${tool.name}`}
-                                          onSelect={() => {
-                                            updateStep(step.id, { config: { ...step.config, tool_name: tool.name } });
-                                            setToolSearchOpen(prev => ({ ...prev, [step.id]: false }));
-                                          }}
-                                          className="text-xs"
-                                        >
-                                          <div className="flex items-center gap-2">
-                                            <span className="text-sm">{tool.icon || '🔧'}</span>
-                                            <span>{normalizeToolName(tool.name, 'agentpress')}</span>
-                                          </div>
-                                          <Check
-                                            className={cn(
-                                              "ml-auto h-3 w-3",
-                                              step.config.tool_name === tool.name ? "opacity-100" : "opacity-0"
-                                            )}
-                                          />
-                                        </CommandItem>
-                                      ))}
-                                    </CommandGroup>
-                                  )}
-                                  {agentTools.mcp_tools.length > 0 && (
-                                    <CommandGroup heading="MCP Tools">
-                                      {agentTools.mcp_tools.map((tool) => (
-                                        <CommandItem
-                                          key={`${tool.server || 'default'}-${tool.name}`}
-                                          value={`${normalizeToolName(tool.name, 'mcp')} ${tool.name} ${tool.server || ''}`}
-                                          onSelect={() => {
-                                            updateStep(step.id, { config: { ...step.config, tool_name: tool.server ? `${tool.server}:${tool.name}` : tool.name } });
-                                            setToolSearchOpen(prev => ({ ...prev, [step.id]: false }));
-                                          }}
-                                          className="text-xs"
-                                        >
-                                          <div className="flex items-center gap-2">
-                                            <span className="text-sm">{tool.icon || '🔧'}</span>
-                                            <span>{normalizeToolName(tool.name, 'mcp')}</span>
-                                          </div>
-                                          <Check
-                                            className={cn(
-                                              "ml-auto h-3 w-3",
-                                              step.config.tool_name === (tool.server ? `${tool.server}:${tool.name}` : tool.name) ? "opacity-100" : "opacity-0"
-                                            )}
-                                          />
-                                        </CommandItem>
-                                      ))}
-                                    </CommandGroup>
-                                  )}
-                                </>
-                              ) : (
-                                <CommandItem disabled className="text-xs">Failed to load tools</CommandItem>
-                              )}
-                            </CommandList>
-                          </Command>
-                        </PopoverContent>
-                      </Popover>
-                    )}
-                  </>
+      <div className="space-y-4">
+        <div className="flex items-center gap-2">
+          {conditionSteps.map((step, index) => {
+            const letter = getConditionLetter(index);
+            const isActive = step.id === activeTabId;
+            const conditionType = step.conditions?.type === 'if' ? 'If' : 
+                                step.conditions?.type === 'elseif' ? 'Else If' :
+                                step.conditions?.type === 'else' ? 'Else' : 'If';
+            return (
+              <button
+                key={step.id}
+                onClick={() => setActiveConditionTab(prev => ({ ...prev, [groupKey]: step.id }))}
+                onKeyDown={(e) => handleKeyDown(e, step)}
+                tabIndex={0}
+                className={cn(
+                  "flex items-center gap-2 px-3 py-2 rounded-md border text-sm font-medium transition-all",
+                  isActive 
+                    ? "bg-primary text-primary-foreground border-primary shadow-sm"
+                    : "bg-background border-border text-foreground hover:bg-accent hover:text-accent-foreground"
                 )}
-
-                <div className="flex items-center gap-1 pt-1">
-                  {!isCondition && (
-                    <>
-                      <Button
-                        variant="ghost"
-                        size="sm"
-                        onClick={() => addStep(undefined, step.id)}
-                        className="h-6 px-2 text-xs hover:bg-primary/10"
-                      >
-                        <Plus className="h-3 w-3" />
-                        Step
-                      </Button>
-                      <Button
-                        variant="ghost"
-                        size="sm"
-                        onClick={() => addCondition(step.id)}
-                        className="h-6 px-2 text-xs"
-                      >
-                        <GitBranch className="h-3 w-3" />
-                        Condition
-                      </Button>
-                    </>
-                  )}
-                  {isCondition && (
-                    <Button
-                      variant="ghost"
-                      size="sm"
-                      onClick={() => addStep(step.id)}
-                      className="h-6 px-2 text-xs hover:bg-primary/10"
-                    >
-                      <Plus className="h-3 w-3" />
-                      Add to branch
-                    </Button>
-                  )}
-                  <Button
-                    variant="ghost"
-                    size="sm"
-                    onClick={() => removeStep(step.id)}
-                    className="h-6 px-2 text-xs text-destructive hover:text-destructive hover:bg-destructive/10 ml-auto"
-                  >
-                    <Trash2 className="h-3 w-3" />
-                  </Button>
-                </div>
+              >
+                <span className="font-mono text-xs">{letter}</span>
+                <span>•</span>
+                <span>{conditionType}</span>
+                {step.hasIssues && (
+                  <AlertTriangle className="h-3 w-3 text-destructive" />
+                )}
+              </button>
+            );
+          })}
+          {!hasElse && (
+            <Button
+              variant="outline"
+              size="sm"
+              onClick={() => addElseCondition(conditionSteps[conditionSteps.length - 1].id)}
+              className="h-9 px-3 border-dashed text-xs"
+            >
+              <Plus className="h-3 w-3 mr-1" />
+              Else If
+            </Button>
+          )}
+          {!hasElse && (
+            <Button
+              variant="outline"
+              size="sm"
+              onClick={() => addFinalElse(conditionSteps[conditionSteps.length - 1].id)}
+              className="h-9 px-3 border-dashed text-xs"
+            >
+              <Plus className="h-3 w-3 mr-1" />
+              Else
+            </Button>
+          )}
+        </div>
+        {activeStep && (
+          <div className="bg-muted/50 rounded-lg p-4 border">
+            {(activeStep.conditions?.type === 'if' || activeStep.conditions?.type === 'elseif') ? (
+              <div className="space-y-3">
+                <Label className="text-sm font-medium">
+                  {activeStep.conditions?.type === 'if' ? 'Condition' : 'Else If Condition'}
+                </Label>
+                <Input
+                  type="text"
+                  value={activeStep.conditions.expression || ''}
+                  onChange={(e) => updateStep(activeStep.id, { 
+                    conditions: { ...activeStep.conditions, expression: e.target.value } 
+                  })}
+                  placeholder="e.g., user asks about pricing"
+                  className="w-full bg-transparent text-sm px-3 py-2 rounded-md"
+                />
               </div>
-            </Card>
-
-            {hasChildren && (
-              <div className="mt-2 relative">
-                <div className="absolute left-4 -top-1 bottom-0 w-px bg-border" />
-                <div className="space-y-2 pt-1">
-                  {step.children!.map((child, index) => 
-                    renderStep(child, depth + 1, index === step.children!.length - 1)
-                  )}
-                </div>
+            ) : (
+              <div className="text-sm text-muted-foreground font-medium">
+                Otherwise (fallback condition)
               </div>
             )}
+            <div className="mt-4 space-y-3">
+              {activeStep.children && activeStep.children.length > 0 && (
+                <>
+                  {activeStep.children.map((child, index) => renderStep(child, index + 1, true, activeStep.id))}
+                </>
+              )}
+              <div className="flex justify-center pt-2">
+                <Button
+                  variant="outline"
+                  size="sm"
+                  onClick={() => addStep(activeStep.id)}
+                  className="border-dashed text-xs"
+                >
+                  <Plus className="h-3 w-3" />
+                  Add step
+                </Button>
+              </div>
+            </div>
           </div>
-        </div>
-
-        {!isLast && depth === 0 && (
-          <div className="absolute left-4 top-10 h-4 w-px bg-border" />
         )}
       </div>
     );
   };
 
+  const renderStep = (step: ConditionalStep, stepNumber: number, isNested: boolean = false, parentId?: string) => {
+    const isCondition = step.type === 'condition';
+    const isSequence = step.type === 'sequence';
+
+    if (isCondition) {
+      return null;
+    }
+
+    return (
+      <div key={step.id} className="group">
+        <div className="bg-card rounded-lg border shadow-sm p-4 hover:shadow-md transition-shadow">
+          <div className="flex items-start gap-4">
+            <div className="flex items-center gap-2 shrink-0">
+              {step.hasIssues && (
+                <AlertTriangle className="h-4 w-4 text-destructive" />
+              )}
+              <div className="w-6 h-6 rounded-full bg-muted flex items-center justify-center text-sm font-medium text-muted-foreground">
+                {stepNumber}
+              </div>
+            </div>
+
+            <div className="flex-1 min-w-0">
+              <div className="flex items-center justify-between mb-3">
+                {isSequence ? (
+                  <div className="flex items-center gap-2">
+                    <div className="w-5 h-5 rounded bg-primary/10 flex items-center justify-center">
+                      <div className="w-2 h-2 rounded-full bg-primary" />
+                    </div>
+                    <span className="text-base font-medium">{step.description}</span>
+                  </div>
+                ) : (
+                  <input
+                    type="text"
+                    value={step.name + ' ' + stepNumber}
+                    onChange={(e) => updateStep(step.id, { name: e.target.value })}
+                    placeholder="Step name"
+                    className="w-full bg-transparent border-0 outline-none text-base font-medium placeholder:text-muted-foreground"
+                  />
+                )}
+              </div>
+              {!isSequence && step.description !== undefined && (
+                <input
+                  type="text"
+                  value={step.description}
+                  onChange={(e) => updateStep(step.id, { description: e.target.value })}
+                  placeholder="Add a description"
+                  className="-mt-2 w-full bg-transparent border-0 outline-none text-sm text-muted-foreground placeholder:text-muted-foreground mb-3"
+                />
+              )}
+              {!isSequence && (
+                <Popover 
+                  open={toolSearchOpen[step.id] || false} 
+                  onOpenChange={(open) => setToolSearchOpen(prev => ({ ...prev, [step.id]: open }))}
+                >
+                  <PopoverTrigger asChild>
+                    <Button
+                      variant="outline"
+                      role="combobox"
+                      aria-expanded={toolSearchOpen[step.id] || false}
+                      className="h-9 w-full justify-between text-sm"
+                    >
+                      {step.config.tool_name ? (
+                        <span className="flex items-center gap-2 text-sm">
+                          {(() => {
+                            const agentpressTool = agentTools?.agentpress_tools.find(t => t.name === step.config.tool_name);
+                            if (agentpressTool) {
+                              return (
+                                <>
+                                  <span>{agentpressTool.icon || '🔧'}</span>
+                                  <span>{normalizeToolName(agentpressTool.name, 'agentpress')}</span>
+                                </>
+                              );
+                            }
+                            const mcpTool = agentTools?.mcp_tools.find(t => `${t.server}:${t.name}` === step.config.tool_name);
+                            if (mcpTool) {
+                              return (
+                                <>
+                                  <span>{mcpTool.icon || '🔧'}</span>
+                                  <span>{normalizeToolName(mcpTool.name, 'mcp')}</span>
+                                </>
+                              );
+                            }
+                            return step.config.tool_name;
+                          })()}
+                        </span>
+                      ) : (
+                        <span className="text-muted-foreground">Select tool (optional)</span>
+                      )}
+                      <ChevronsUpDown className="ml-2 h-4 w-4 shrink-0 opacity-50" />
+                    </Button>
+                  </PopoverTrigger>
+                  <PopoverContent className="w-[320px] p-0" align="start">
+                    <Command>
+                      <CommandInput placeholder="Search tools..." className="h-9" />
+                      <CommandEmpty>No tools found.</CommandEmpty>
+                      <CommandList>
+                        {isLoadingTools ? (
+                          <CommandItem disabled>Loading tools...</CommandItem>
+                        ) : agentTools ? (
+                          <>
+                            {agentTools.agentpress_tools.filter(tool => tool.enabled).length > 0 && (
+                              <CommandGroup heading="AgentPress Tools">
+                                {agentTools.agentpress_tools.filter(tool => tool.enabled).map((tool) => (
+                                  <CommandItem
+                                    key={tool.name}
+                                    value={`${normalizeToolName(tool.name, 'agentpress')} ${tool.name}`}
+                                    onSelect={() => {
+                                      updateStep(step.id, { config: { ...step.config, tool_name: tool.name } });
+                                      setToolSearchOpen(prev => ({ ...prev, [step.id]: false }));
+                                    }}
+                                  >
+                                    <div className="flex items-center gap-2">
+                                      <span>{tool.icon || '🔧'}</span>
+                                      <span>{normalizeToolName(tool.name, 'agentpress')}</span>
+                                    </div>
+                                    <Check
+                                      className={cn(
+                                        "ml-auto h-4 w-4",
+                                        step.config.tool_name === tool.name ? "opacity-100" : "opacity-0"
+                                      )}
+                                    />
+                                  </CommandItem>
+                                ))}
+                              </CommandGroup>
+                            )}
+                            {agentTools.mcp_tools.length > 0 && (
+                              <CommandGroup heading="MCP Tools">
+                                {agentTools.mcp_tools.map((tool) => (
+                                  <CommandItem
+                                    key={`${tool.server || 'default'}-${tool.name}`}
+                                    value={`${normalizeToolName(tool.name, 'mcp')} ${tool.name} ${tool.server || ''}`}
+                                    onSelect={() => {
+                                      updateStep(step.id, { config: { ...step.config, tool_name: tool.server ? `${tool.server}:${tool.name}` : tool.name } });
+                                      setToolSearchOpen(prev => ({ ...prev, [step.id]: false }));
+                                    }}
+                                  >
+                                    <div className="flex items-center gap-2">
+                                      <span>{tool.icon || '🔧'}</span>
+                                      <span>{normalizeToolName(tool.name, 'mcp')}</span>
+                                    </div>
+                                    <Check
+                                      className={cn(
+                                        "ml-auto h-4 w-4",
+                                        step.config.tool_name === (tool.server ? `${tool.server}:${tool.name}` : tool.name) ? "opacity-100" : "opacity-0"
+                                      )}
+                                    />
+                                  </CommandItem>
+                                ))}
+                              </CommandGroup>
+                            )}
+                          </>
+                        ) : (
+                          <CommandItem disabled>Failed to load tools</CommandItem>
+                        )}
+                      </CommandList>
+                    </Command>
+                  </PopoverContent>
+                </Popover>
+              )}
+              {step.children && step.children.length > 0 && (
+                <div className="mt-4 space-y-4">
+                  {step.children.map((child, index) => renderStep(child, index + 1, true, step.id))}
+                </div>
+              )}
+            </div>
+            <Popover>
+              <PopoverTrigger asChild>
+                <Button variant="ghost" size="sm" className="h-8 w-8 p-0 opacity-0 group-hover:opacity-100 transition-opacity">
+                  <MoreHorizontal className="h-4 w-4" />
+                </Button>
+              </PopoverTrigger>
+              <PopoverContent className="w-48 p-1" align="end">
+                <Button
+                  variant="ghost"
+                  size="sm"
+                  onClick={() => removeStep(step.id)}
+                  className="w-full justify-start text-destructive hover:text-destructive hover:bg-destructive/10"
+                >
+                  <Trash2 className="h-4 w-4 mr-2" />
+                  Delete step
+                </Button>
+              </PopoverContent>
+            </Popover>
+          </div>
+        </div>
+      </div>
+    );
+  };
+
+  const renderSteps = () => {
+    const result: React.ReactNode[] = [];
+    let stepCounter = 0;
+    let i = 0;
+
+    while (i < steps.length) {
+      const step = steps[i];
+      if (step.type === 'condition') {
+        const conditionGroup: ConditionalStep[] = [];
+        while (i < steps.length && steps[i].type === 'condition') {
+          conditionGroup.push(steps[i]);
+          i++;
+        }
+        stepCounter++;
+        result.push(
+          <div key={conditionGroup[0].id} className="bg-card rounded-lg border shadow-sm p-4 hover:shadow-md transition-shadow">
+            <div className="flex items-start gap-4">
+              <div className="flex items-center gap-2 shrink-0">
+                <div className="w-6 h-6 rounded-full bg-muted flex items-center justify-center text-sm font-medium text-muted-foreground">
+                  {stepCounter}
+                </div>
+              </div>
+              <div className="flex-1">
+                <div className="text-base font-medium mb-4">Add rule</div>
+                {renderConditionTabs(conditionGroup, conditionGroup[0].id)}
+              </div>
+            </div>
+          </div>
+        );
+      } else {
+        stepCounter++;
+        result.push(renderStep(step, stepCounter, false));
+        i++;
+      }
+    }
+
+    return result;
+  };
+
   return (
-    <div className="space-y-4">
+    <div className="space-y-6 max-w-4xl">
       {steps.length === 0 ? (
         <div className="text-center py-16">
-          <div className="w-14 h-14 bg-primary/10 rounded-2xl flex items-center justify-center mx-auto mb-4">
-            <GitBranch className="h-7 w-7 text-primary" />
+          <div className="w-16 h-16 bg-muted rounded-2xl flex items-center justify-center mx-auto mb-4">
+            <Plus className="h-8 w-8 text-muted-foreground" />
           </div>
-          <h3 className="text-base font-semibold mb-2">Start building your workflow</h3>
-          <p className="text-sm text-muted-foreground mb-6 max-w-sm mx-auto">
-            Add steps and conditions to create a smart workflow that adapts to different scenarios
+          <h3 className="text-lg font-semibold mb-2">Start building your workflow</h3>
+          <p className="text-muted-foreground mb-6 max-w-md mx-auto">
+            Add steps and conditions to create a smart workflow that adapts to different scenarios.
           </p>
           <Button 
-            variant="outline" 
             onClick={() => addStep()}
-            className="border-dashed hover:border-solid"
           >
             <Plus className="h-4 w-4 mr-2" />
-            Add first step
+            Add step
           </Button>
         </div>
       ) : (
-        <div className="space-y-2">
-          {steps.map((step, index) => renderStep(step, 0, index === steps.length - 1))}
-          <div className="pl-12 pt-2">
-            <Button 
-              variant="outline" 
-              onClick={() => addStep()}
-              className="h-8 border-dashed text-xs hover:border-solid"
-              size="sm"
-            >
-              <Plus className="h-3 w-3" />
-              Add step
-            </Button>
+        <div className="space-y-6">
+          {renderSteps()}
+          <div className="flex justify-center pt-4">
+            <div className="flex gap-3">
+              <Button 
+                variant="outline"
+                onClick={() => addStep()}
+                className="border-dashed"
+              >
+                <Plus className="h-4 w-4" />
+                Add step
+              </Button>
+              
+              <Button 
+                variant="outline"
+                onClick={() => addCondition(steps[steps.length - 1]?.id || '')}
+                className="border-dashed"
+              >
+                <Plus className="h-4 w-4" />
+                Add rule
+              </Button>
+            </div>
           </div>
         </div>
       )}
     </div>
   );
-} 
+}
